# inventory-service/inventory-infrastructure-rest/src/main/resources/static/openapi.yaml
openapi: 3.1.1
info:
  title: Swagger Inventory Service
  description: |-
    Bienvenido a la documentación oficial del microservicio 'Payment Service'. Aquí podrás explorar e interactuar con los 
    diversos recursos disponibles través de su API. 
    Esta documentación proporciona los detalles técnicos necesarios para utilizar las funcionalidades ofrecidas por el 
    servicio de manera eficiente.

    Algunos enlaces útiles:
      - [El repositorio de Inventory Service](https://github.com/ByLuisen/my-k8s-project/tree/main/apps/order-service)
      - [La definición de la API para Inventory Service](https://github.com/ByLuisen/my-k8s-project/blob/main/apps/order-service/src/main/resources/static/openapi.yaml)
  version: 1.0.12
servers:
  - url: /api
...
paths:
...
  /v1/me/offers/{offerId}:
    put:
      tags:
        - Offers
      summary: Actualiza los detalles de tu oferta
      description: Actualiza los detalles de tu oferta
      operationId: updateOffer
      parameters:
        - name: offerId
          in: path
          description: ID de la oferta
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Información necesaria para actualizar tu oferta
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfferRequestDto'
        required: true
      responses:
        '200':
          description: Operción exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/OfferResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: Oferta no encontrada
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [supplier]
      x-roles-required:
        - SUPPLIER
    delete:
      tags:
        - Offers
      summary: Elimina una oferta
      description: Elimina una oferta
      operationId: deleteOffer
      parameters:
        - name: offerId
          in: path
          description: ID de la oferta
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Operación exitosa
        '400':
          description: Solicitud incorrecta
        '404':
          description: Oferta no encontrada
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [supplier]
      x-roles-required:
        - SUPPLIER
  /v1/me/offers:
    post:
      tags:
        - Offers
      summary: Crea una nueva oferta
      description: Crea una nueva oferta
      operationId: createOffer
      requestBody:
        description: Información necesaria para crear una nueva oferta
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfferRequestDto'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/OfferResponse'
        '400':
          description: Solicitud incorrecta
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [supplier]
      x-roles-required:
        - SUPPLIER
  /v1/suppliers/{supplierId}/offers:
    get:
      tags:
        - Offers
      summary: Lista las diferentes ofertas que tiene un proveedor
      description: Lista las diferentes ofertas que tiene un proveedor
      operationId: getSupplierOffers
      parameters:
        - name: supplierId
          in: path
          descrption: ID del proveedor
          required: true
          schema:
            type: string
            format: uuid
        - name: offerBySupplierIdRequestDto
          in: query
          description: Filtros disponibles
          required: true
          schema:
            $ref: '#/components/schemas/OfferBySupplierIdRequestDto'
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: >-
                  #/components/schemas/PaginatedResultResponseSupplierOfferReadResponse
        '400':
          description: Solicitud incorrecta
        '404':
          description: Ofertas del proveedor no encontradas
        default:
          description: Error inesperado
  /v1/products/{productId}/offers:
    get:
      tags:
        - Offers
      summary: Lista las ofertas que tiene un producto
      description: Lista las ofertas que tiene un producto
      operationId: getProductOffers
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductOfferReadResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: Ofertas del producto no encontradas
        default:
          description: Error inesperado
...
  /v1/departments/{departmentId}:
    put:
      tags:
        - Departments
      summary: Actualiza la información de un departamento
      description: Actualiza la información de un departamento
      operationId: updateDepartment
      parameters:
        - name: departmentId
          in: path
          description: ID del departamento
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Información necesaria para actualizar un departamento
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequestDto'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: Departamento no encontrado
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [admin]
      x-roles-required:
        - ADMIN
    delete:
      tags:
        - Departments
      summary: Elimina un departamento
      description: Elimina un departamento
      operationId: deleteDepartmentById
      parameters:
        - name: departmentId
          in: path
          description: ID del departamento
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Operación exitosa
        '400':
          description: Solicitud incorrecta
        '404':
          description: Departamento no encontrado
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [admin]
      x-roles-required:
        - ADMIN
  /v1/departments:
    get:
      tags:
        - Departments
      summary: Lista nuestros diferentes departamentos
      description: Lista nuestros diferetes departamentos
      operationId: findAllDepartments
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentResponse'
        '404':
          description: Departamentos no encontrados
        default:
          description: Error inesperado
    post:
      tags:
        - Departments
      summary: Crea un nuevo departamento
      description: Crea un nuevo departamento
      operationId: saveDepartment
      requestBody:
        description: Información necesaria para crear un nuevo departamento
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequestDto'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '400':
          description: Solicitud incorrecta
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [amdin]
      x-roles-required:
        - ADMIN
...
  /v1/images:
    get:
      tags:
        - Images
      summary: Lista las imágenes de un producto
      description: Lista las imágenes de un producto
      operationId: getProductImages
      parameters:
        - name: productId
          in: query
          description: ID del producto
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: Imágenes del producto no encontradas
        default:
          description: Error inesperado
    post:
      tags:
        - Images
      summary: Guarda una imagen
      description: Guarda una imagen
      operationId: saveImage
      requestBody:
        description: Información necesaria para guardar una imagen
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageRequestDto'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: La información proporcionada no existe
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [admin]
      x-roles-required:
        - ADMIN
  /v1/images/{imageId}:
    delete:
      tags:
        - Images
      summary: Elimina una imagen
      description: Elimina una imagen
      operationId: deleteImageById
      parameters:
        - name: imageId
          in: path
          description: ID de la imagen
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Operación exitosa
        '400':
          description: Solicitud incorrecta
        '404':
          description: Imagen no encontrada
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [admin]
      x-roles-required:
        - ADMIN
...
components:
  schemas:
...
    UpdateOfferRequestDto:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        price:
          type: number
          minimum: 0.01
          examples: [79.99]
        stock:
          type: integer
          format: int32
          examples: [10]
      required:
        - price
        - productId
        - stock
    OfferResponse:
      type: object
      properties:
        offerId:
          type: string
          format: uuid
        supplierId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        price:
          type: number
          examples: [79.99]
        stock:
          type: integer
          format: int32
          examples: [10]
...
    UpdateDepartmentRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 3
          pattern: ^[A-Za-z0-9][A-Za-z0-9 .-]*[A-Za-z0-9]$
          examples: ['Electronics']
      required:
        - name
    DepartmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          examples: ['Electronics']
...
    CreateOfferRequestDto:
      type: object
      properties:
        productId:
          type: string
        price:
          type: number
          minimum: 0.01
          examples: [79.99]
        stock:
          type: integer
          format: int32
          examples: [10]
      required:
        - price
        - productId
        - stock
...
    CreateImageRequestDto:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        imageUrl:
          type: string
          pattern: ^(https?://.*\.(?:png|jpg|jpeg|gif|bmp|webp))$
          examples: ['https://example.com/images/productos/camiseta-deportiva.jpg']
      required:
        - imageUrl
        - productId
    CreateDepartmentRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 3
          pattern: ^[A-Za-z0-9][A-Za-z0-9 .-]*[A-Za-z0-9]$
          examples: ['Electronics']
      required:
        - name
...
    OfferBySupplierIdRequestDto:
      type: object
      properties:
        size:
          type: integer
          format: int32
          minimum: 1
          examples: [10]
        page:
          type: integer
          format: int32
          examples: [0]
      required:
        - page
        - size
    PaginatedResultResponseSupplierOfferReadResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SupplierOfferReadResponse'
        count:
          type: integer
          format: int64
          examples: [1]
    SupplierOfferProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          examples: ['SAM-001-EL']
        name:
          type: string
          examples: ['Samsung Galaxy S23']
    SupplierOfferReadResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/SupplierOfferProductResponse'
        stock:
          type: integer
          format: int32
          examples: [10]
        price:
          type: number
          examples: [79.99]
...
    ProductOfferSupplierResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          examples: ['Bambi']
    ProductOfferReadResponse:
      type: object
      properties:
        supplier:
          $ref: '#/components/schemas/ProductOfferSupplierResponse'
        stock:
          type: integer
          format: int32
          examples: [10]
        price:
          type: number
          examples: [79.99]
...
    ImageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        imageUrl:
          type: string
          examples: ['https://example.com/images/productos/camiseta-deportiva.jpg']
...
  securitySchemes:
    BambiOAuth2:
      type: oauth2
      flows:
        authorizationCode:
#          authorizationUrl: http://localhost:8081/realms/bambi/protocol/openid-connect/auth
#          tokenUrl: http://localhost:8081/realms/bambi/protocol/openid-connect/token
          authorizationUrl: http://keycloak.local/realms/bambi/protocol/openid-connect/auth
          tokenUrl: http://keycloak.local/realms/bambi/protocol/openid-connect/token
          scopes:
            - admin: Acceso administrativo
            - supplier: Acceso proveedor